Only truncate if size actually is set to something smaller

+++ src/jext2/RegularInode.java
@@ -27,8 +27,10 @@ public class RegularInode extends DataInode {
      * @param   size    new size
      */
     public void setSizeAndTruncate(long size) throws IOException {
+        long oldSize = getSize();
         setSize(size);
-        accessData().truncate();
+        if (oldSize > size)
+            accessData().truncate();
     }
 
     /**
